refactor(home): Error message cart empty
Error message cart empty

see #374

+++ app/Http/Controllers/OrderController.php
@@ -15,11 +15,21 @@ class OrderController extends Controller
 {
     public function show()
     {
+        if (Cart::count() == 0) {
+            $error = "Cart is empty! Please add products.";
+            $carts = Cart::content();
+            $countries = Country::pluck('name', 'id');
+            return view('carts.cart')
+                ->with('carts', $carts)
+                ->with('error', $error)
+                ->with('countries', $countries);
+        } else {
         $carts = Cart::content();
         $countries = Country::pluck('name', 'id');
         return view('orders.order')
             ->with('carts', $carts)
             ->with('countries', $countries);
+        }
     }
 
     public function store(Request $request)
@@ -34,20 +44,30 @@ class OrderController extends Controller
             'country_id' => 'required|numeric|exists:countries,id',
         ]);
 
-        $order_id = mt_rand();
-        $data = $request->all();
-        $data['id'] = $order_id;
-        unset($data['_token']);
-        if (!Auth::guest()) {
-            $data['user_id'] = Auth::user()->id;
+        if (Cart::count() == 0) {
+            $error = "Cart is empty! Please add products.";
+            $carts = Cart::content();
+            $countries = Country::pluck('name', 'id');
+            return view('carts.cart')
+                ->with('carts', $carts)
+                ->with('error', $error)
+                ->with('countries', $countries);
+        } else {
+            $order_id = mt_rand();
+            $data = $request->all();
+            $data['id'] = $order_id;
+            unset($data['_token']);
+            if (!Auth::guest()) {
+                $data['user_id'] = Auth::user()->id;
+            }
+            Order::insert($data);
+            OrderProduct::store($order_id);
+            $subtotal = Cart::subtotal();
+            Cart::destroy();
+            $order = Order::findOrFail($order_id);
+            return view('orders.orderInformation')
+                ->with('order', $order)
+                ->with('subtotal', $subtotal);
         }
-        Order::insert($data);
-        OrderProduct::store($order_id);
-        $subtotal = Cart::subtotal();
-        Cart::destroy();
-        $order = Order::findOrFail($order_id);
-        return view('orders.orderInformation')
-            ->with('order', $order)
-            ->with('subtotal', $subtotal);
     }
 }
